Project 3 Update 2
Still working on part a.  Things run, but the Verlet algorithm (or the
velocity/acceleration being fed into the Verlet algorithm, which is
perhaps more likely) is not putting out reasonable things.  Also, for
some reason the plot isn't working of the position and velocity, and I
have no idea why?